Initial Update
This is the initial update of this project.  The two csv files are the
two data sets used in this app.

At this point the app runs with no errors.  It has a title “Wine
Quality” and a sidebar with 11 numerical inputs.  There is an output
(quality) but is is not displayed on the app yet